proj/requirements.txt
django>=3.0,<3.1
Pillow
uWSGI
requests
django-cas-ng
django-mssql-backend
djangorestframework
django-cors-headers
django-simple-history
django-debug-toolbar
coreapi
Pygments
Markdown
packages/django-api-key-auth-0.3.tar.gz